refactor out stackexchange_results
