+++ src/sentry/static/sentry/app/components/events/stackExchange/styles.jsx
@@ -1,0 +1,16 @@
+import styled from 'react-emotion';
+import {Flex} from 'grid-emotion';
+
+export const Sticky = styled('div')`
+  position: sticky;
+  z-index: ${p => p.theme.zIndex.header};
+  top: -1px;
+`;
+
+export const StyledFlex = styled(Flex)`
+  align-items: center;
+  background: ${p => p.theme.offWhite};
+  border-bottom: 1px solid ${p => p.theme.borderDark};
+  border-radius: ${p => p.theme.borderRadius} ${p => p.theme.borderRadius} 0 0;
+  margin-bottom: -1px;
+`;
